[Feat] Team page 진입 시 팀유무에 따른 라우팅 설정
